+++ tests/testthat/test-row-col-fns.R
@@ -40,11 +40,11 @@ test_that('set_properties works with cell functions', {
   ht <- set_font(ht, odds(), 1:2, 'palatino')
   expect_equivalent(font(ht), matrix(c('palatino', 'times'), 4, 2))
   ht <- hux(a = 1:4, b = 1:4)
-  ht <- set_font(ht, , evens(), 'times')
-  ht <- set_font(ht, , odds(), 'palatino')
+  ht <- set_font(ht, every(1), evens(), 'times')
+  ht <- set_font(ht, every(1), odds(), 'palatino')
   expect_equivalent(font(ht), matrix(c('palatino', 'times'), 4, 2, byrow = TRUE))
   ht <- hux(a = 1:4, b = 1:4)
-  ht <- set_font(ht, every(3, from = 1), ,'times')
+  ht <- set_font(ht, every(3, from = 1), every(1),'times')
   expect_equivalent(font(ht), matrix(c('times', NA, NA, 'times'), 4, 2))
 })
 
@@ -61,7 +61,7 @@ test_that('set_properties works with row and column functions', {
 
 test_that('set_properties works with column ranges', {
   ht <- hux(a = 1:4, b = 1:4, c = 1:4, d = 1:4)
-  ht <- set_font(ht, , b:d, 'times')
+  ht <- set_font(ht, every(1), b:d, 'times')
   expect_equivalent(font(ht), matrix(c(NA, 'times', 'times', 'times'), 4 , 4, byrow = TRUE))
 })
 
@@ -71,13 +71,18 @@ test_that('set_properties works with expressions in ht context', {
   expect_equivalent(font(ht), matrix(c(NA, 'times', 'times', NA), 4 , 2))
 })
 
-test_that('set_properties works with expressions in ht context', {
+test_that('set_* works with byrow', {
   ht <- hux(a = 1:2, b = 1:2)
   ht <- set_font(ht, 1:2, 1:2, c('times', 'palatino'), byrow = TRUE)
   expect_equivalent(font(ht), matrix(c('times', 'times', 'palatino', 'palatino'), 2, 2))
+
   ht <- hux(a = 1:2, b = 1:2, d = 1:2)
   ht <- set_font(ht, 2, 1:2, c('times', 'palatino'), byrow = TRUE)
   expect_equivalent(font(ht), matrix(c(NA, 'times', NA, 'palatino', NA, NA), 2, 3))
+
+  ht <- hux(a = 1:2, b = 1:2)
+  ht <- set_font(ht, c('times', 'palatino'), byrow = TRUE)
+  expect_equivalent(font(ht), matrix(c('times', 'times', 'palatino', 'palatino'), 2, 2))
 })
 
 test_that('where() works as expected', {
@@ -85,10 +90,32 @@ test_that('where() works as expected', {
   expect_equivalent(where(dfr == 3), matrix(c(3, 1, 1, 3), 2, 2))
 })
 
-test_that('where() works with set_properties', {
+test_that('where() works with set_*', {
   ht <- hux(a = c(1, 0), b = c(0, 1))
   ht <- set_font(ht, where(ht > 0), 'times')
   expect_equivalent(font(ht), matrix(c('times', NA, NA, 'times'), 2, 2))
+})
+
+test_that('set_* works when col is missing', {
+  ht <- hux(a = c(1, 0), b = c(0, 1))
+  ht2 <- set_font(ht, where(ht > 0), value = 'times')
+  expect_equivalent(font(ht2), matrix(c('times', NA, NA, 'times'), 2, 2))
+  ht3 <- set_font(ht, where(ht > 0), 'times')
+  expect_equivalent(font(ht3), matrix(c('times', NA, NA, 'times'), 2, 2))
+  expect_error(set_font(ht, 1:3, 'times')) # no cols!
+  expect_error(set_font(ht, where(ht > 0), 'times', byrow = TRUE)) # hard to interpret
+})
+
+test_that('set_* works when row and col are missing', {
+  ht <- hux(a = c(1, 0), b = c(0, 1))
+  ht2 <- set_font(ht, 'times')
+  expect_equivalent(font(ht2), matrix('times', 2, 2))
+  ht3 <- set_font(ht, value = 'times')
+  expect_equivalent(font(ht3), matrix('times', 2, 2))
+  ht4 <- set_font(ht, value = c('times', 'arial'), byrow = TRUE)
+  expect_equivalent(font(ht4), matrix(c('times', 'times', 'arial', 'arial'), 2, 2))
+  ht5 <- set_font(ht, c('times', 'arial'), byrow = TRUE)
+  expect_equivalent(font(ht5), matrix(c('times', 'times', 'arial', 'arial'), 2, 2))
 })
 
 test_that('is_a_number() works as expected', {