+++ NEWS.md
@@ -1,6 +1,7 @@
 # huxtable 0.1.1
 
 * Avoid `strrep`; try to pass CHECK where pandoc is not available
+* Try to make vignettes build without pandoc
 
 # huxtable 0.1.0