+++ NEWS.md
@@ -4,9 +4,10 @@
 * New set_* interface allowing column ranges, expressions a la `subset`, and filling in values by row.
 * New function `set_cell_properties` to set multiple properties on cells.
 * `evens()`, `odds()`, `every()`, `final()`, `where()`: convenience functions to select rows, columns and cells.
-* Exported function `guess_knitr_output_format`.
 * Export to Word/Powerpoint via `ReporteRs`.
 * Huxtable now supports dplyr verbs like `filter` and `select`.
+* Exported function `guess_knitr_output_format`.
+* Ability to set border colors.
 * Prevent overlapping row/colspans.
 
 # huxtable 0.1.0