+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * New `huxreg` function to convert a list of models to a huxtable.
 * New set_* interface allowing column ranges, expressions a la `subset`, and filling in values by row.
+* Replacement methods `$<-`, `[<-` and `[[<-` now work better.
 * New function `set_cell_properties` to set multiple properties on cells.
 * `evens()`, `odds()`, `every()`, `final()`, `where()`: convenience functions to select rows, columns and cells.
 * Export to Word/Powerpoint via `ReporteRs`.